refactor(scrollview): extract pointer logic out of ScrollView
